update test cases and shift session to generated-outputs folder
