+++ tests/test_runner.py
@@ -1,4 +1,5 @@
 import asyncio
+from moonshot.src.api.api_run import api_get_all_run
 from moonshot.src.api.api_runner import (
     api_create_runner, api_delete_runner, api_get_all_runner, 
     api_get_all_runner_name, api_load_runner, api_read_runner
@@ -187,6 +188,28 @@ def test_run_runner_api():
     print("=" * 100, "\nTest listing all runner name")
     test_get_all_runner_name()
 
+    # Test Run API
+    test_run_api()
+
     # Delete runner
     print("=" * 100, "\nTest deleting runners")
     test_delete_runner(runner_id)
+
+# ------------------------------------------------------------------------------
+# Run APIs Test
+# ------------------------------------------------------------------------------
+def test_get_all_run(runner_id: str):
+    print(api_get_all_run(runner_id))
+
+def test_run_api():
+    # Get run info
+    print("=" * 100, "\nGetting run information for my-new-runner")
+    test_get_all_run("my-new-runner")
+
+    # Get run info
+    print("=" * 100, "\nGetting run information for no-runner")
+    test_get_all_run("no-runner")
+
+    # Get run info
+    print("=" * 100, "\nGetting run information for empty string")
+    test_get_all_run("")
